mkrelease was broken on clean working dir by stash stuff

+++ mkrelease.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
 set -e
+set -x
 
 # Get mod info from info.json
 mod_name=$(perl -ne 'if (/"name":\s+"([^"]+)"/) { print("$1\n"); }' < src/info.json)
@@ -9,10 +10,21 @@ mod_version=$(perl -ne 'if (/"version":\s+"([^"]+)"/) { print("$1\n"); }' < src/
 mod_dir_name="${mod_name}_${mod_version}"
 mod_zip_name="${mod_dir_name}.zip"
 
-# Create the zip
-# Create a stash so that we get local changes in the working dir
-stash_name=$(git stash create)
-git stash store $stash_name # Store it so we can later call drop on it
-git archive -o "releases/$mod_zip_name" --prefix "$mod_dir_name/" $stash_name:src
-# Cleanup; really we should put this in an exit hook or something in case the above errors
-git stash drop stash@{0}
+function archive() {
+	ref=$1
+	# Create the zip
+	git archive -o "releases/$mod_zip_name" --prefix "$mod_dir_name/" "$ref":src
+}
+
+if git diff-index --quiet HEAD --; then
+	# No changes
+	archive HEAD
+else
+	# There are working dir changes
+	# Create a stash so that we get local changes in the working dir
+	stash_name=$(git stash create)
+	git stash store $stash_name # Store it so we can later call drop on it
+	archive $stash_name
+	# Cleanup; really we should put this in an exit hook or something in case the above errors
+	git stash drop stash@{0}
+fi
